fix: review fixes
move code from index.ts into separate files
use ternary operator
don't explicitely initialize to undefined
remove leftover lint option
