+++ package.json
@@ -25,7 +25,7 @@
     "lint": "eslint src/",
     "format": "prettier -c src/",
     "format:fix": "prettier --write src/",
-    "prepush": "yarn test && yarn lint && yarn format"
+    "prepush": "yarn test --clear-cache && yarn test && yarn lint && yarn format"
   },
   "devDependencies": {
     "@types/jest": "^25.2.1",
