+++ src/language/lexer/sublexer/default/index.ts
@@ -1,32 +1,11 @@
-import { ParseResult } from '../../sublexer';
-import { DefaultSublexerTokenData } from '../../token';
-import {
-  tryParseComment,
-  tryParseIdentifier,
-  tryParseLiteral,
-  tryParseOperator,
-  tryParseSeparator,
-} from './rules';
-import { tryParseStringLiteral } from './string';
-
 export {
   tryParseComment,
   tryParseIdentifier,
   tryParseLiteral,
   tryParseOperator,
   tryParseSeparator,
-  tryParseStringLiteral,
-};
+} from './rules';
 
-export function tryParseDefault(
-  slice: string
-): ParseResult<DefaultSublexerTokenData> {
-  return (
-    tryParseSeparator(slice) ??
-    tryParseOperator(slice) ??
-    tryParseLiteral(slice) ??
-    tryParseStringLiteral(slice) ??
-    tryParseIdentifier(slice) ??
-    tryParseComment(slice)
-  );
-}
+export { tryParseStringLiteral } from './string';
+
+export { tryParseDefault } from './glue';
