+++ src/language/lexer/sublexer/result.ts
